fix(toolbar): cut the accidental LemonToast and LemonTable eager chains
Continues .agents/toolbar-migration.md (Step 7's eager-set audit, traced
with a metafile BFS from the toolbar entry):

- LemonToast statically imported incidentStatusLogic just for the current
  status string and the status-page URL, dragging the help-menu /
  superpowers / MCP-hint graph (including eventUsageLogic and the 224 KB
  taxonomy JSON) into every bundle that can show a toast. Those two live
  in a new leaf, lib/components/HelpMenu/incidentStatus.ts; the logic
  imports and re-exports it.
- LemonTable imported determineColumnKey/getStickyColumnInfo from
  columnUtils.tsx, whose column factories render TZLabel (and through it
  eventUsageLogic -> taxonomy). The pure layout helpers move to
  columnLayoutUtils.ts; columnUtils re-exports them for its importers.

Toolbar eager JS 1.88 -> 1.54 MB (-18%), budget ratcheted 2.1 -> 1.7 MB.
eventUsageLogic leaves the toolbar graph entirely (-25 files), deleting
its last baseline edge (11 -> 10). Both cuts also slim the main app and
exporter entries, which share these chains.

Deliberately kept eager: posthog-js via toolbarPosthogJS (toolbar
telemetry), webVitalsToolbarLogic background collection, and the
per-feature logic connects in toolbarLogic (menu loading spinners).
